pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tfq0seo"
version = "2.2.0"
description = "Fast SEO analysis tool with Lighthouse-quality reports"
authors = [{name = "TFQ0 SEO Team"}]
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Internet :: WWW/HTTP :: Site Management",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "aiohttp>=3.8.0",
    "beautifulsoup4>=4.11.0",
    "click>=8.1.0",
    "jinja2>=3.1.0",
    "rich>=12.6.0",
    "textstat>=0.7.3",
    "validators>=0.20.0",
    "pyyaml>=6.0",
    "python-dateutil>=2.8.0",
    "urllib3>=1.26.0",
    "psutil>=5.9.0",
]

[project.optional-dependencies]
full = [
    "lxml>=4.9.0",
    "openpyxl>=3.0.0",
    "pandas>=1.5.0",
]

[project.scripts]
tfq0seo = "tfq0seo.cli:main"

[tool.setuptools.packages.find]
include = ["tfq0seo*"]

[tool.setuptools.package-data]
tfq0seo = ["templates/*.html"]
